requirements.txt
discord.py
python-dotenv
python-aternos
moviepy